[WARP-SPEED] SIC-2475 storage engine file API use foreign functions (instead of JNI)

+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/main/java/io/trino/plugin/warp/storage/read/StorageReader.java
@@ -148,7 +148,7 @@ public class StorageReader
     void close()
     {
         if (storageCollectorArgs != null) {
-            storageEngine.fileClose(storageCollectorArgs.collectTxArgs().fileCookie()[FILE_COOKIE_PARAMS_FD.ordinal()]);
+            storageEngine.fileClose((int) storageCollectorArgs.collectTxArgs().fileCookie()[FILE_COOKIE_PARAMS_FD.ordinal()]);
         }
     }
